requirements.txt
# 核心依赖
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# 强化学习框架
stable-baselines3>=1.6.0
gym>=0.21.0
gymnasium>=0.26.0

# 深度学习框架
torch>=1.9.0
torchvision>=0.10.0

# 机器学习工具
scikit-learn>=1.0.0
xgboost>=1.5.0

# 数据处理
pyyaml>=6.0
h5py>=3.1.0
tables>=3.6.0

# 可视化
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
bokeh>=2.4.0

# 可解释性
shap>=0.40.0
lime>=0.2.0

# Web框架
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0

# 数据库
sqlalchemy>=1.4.0
sqlite3

# 配置管理
hydra-core>=1.1.0
omegaconf>=2.1.0

# 日志和监控
tensorboard>=2.7.0
wandb>=0.12.0

# 工具库
tqdm>=4.62.0
click>=8.0.0
rich>=10.0.0

# 测试
pytest>=6.2.0
pytest-cov>=2.12.0

# 代码质量
black>=21.0.0
flake8>=3.9.0
mypy>=0.910

# 文档生成
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0

# 其他工具
python-dateutil>=2.8.0
pytz>=2021.1
requests>=2.25.0

